+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.4] - 2019-06-22
+### Added
+- User domain model
+- PhpUnit tests
+
 ## [0.0.3] - 2019-06-22
 ### Added
 - RAW Entities and Repositories